+++ packages/core-js/features/string/index.js
@@ -1,5 +1,7 @@
 var parent = require('../../stable/string');
+// TODO: remove from `core-js@4`
 require('../../modules/esnext.string.at');
+require('../../modules/esnext.string.cooked');
 // TODO: disabled by default because of the conflict with another proposal
 // require('../../modules/esnext.string.at-alternative');
 require('../../modules/esnext.string.code-points');