+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Changelog
 ##### Unreleased
 - Added `Iterator.prototype.toAsync` method from [the iterator helpers stage 2 proposal](https://github.com/tc39/proposal-iterator-helpers)
+- Added [`String.cooked` stage 1 proposal](https://github.com/tc39/proposal-string-cooked):
+  - `String.cooked`
 - Added [`Function.prototype.unThis` stage 0 proposal](https://github.com/js-choi/proposal-function-un-this)
 - Added [`Function.{ isCallable, isConstructor }` stage 0 proposal](https://github.com/caitp/TC39-Proposals/blob/trunk/tc39-reflect-isconstructor-iscallable.md):
   - `Function.isCallable`